requirements.txt
onnxruntime-directml==1.16.3
opencv-python==4.8.1.78
pyttsx3==2.90
numpy==1.24.3
Pillow==10.0.1
